requirements.txt
beautifulsoup4
faiss-cpu
google-api-python-client
google-auth
google-auth-oauthlib
groq
numpy
openai
python-dotenv
sqlalchemy
mysql-connector-python
fastapi
uvicorn
pyjwt
passlib[bcrypt]
python-multipart
email-validator
